002_compose/app/requirements.txt
flask==2.0.0
psycopg2-binary==2.8.6
